issue #20 - Add function to get book

+++ database/main/tables/3_books.sql
@@ -1,11 +1,11 @@
 create table if not exists main.books(
-  id          bigserial   primary key,
-  title       text        not null check(title != ''),
-  price       int         not null,
-  count       int         not null,
-  creator_id  bigint      not null references main.users(id),
+  id          bigserial     primary key,
+  title       text          not null check(title != ''),
+  price       decimal(10,2) not null,
+  count       int           not null,
+  creator_id  bigint        not null references main.users(id),
   description text,
-  created_at  timestamptz default now(),
+  created_at  timestamptz   default now(),
   updated_at  timestamptz
 );
